Move model reader into shared project. Update UWP demo to include a OIT demo
