+++ Source/Examples/UWP/SimpleDemo/SimpleDemo.csproj
@@ -114,6 +114,7 @@
       <DependentUpon>MainPage.xaml</DependentUpon>
     </Compile>
     <Compile Include="MainPageViewModel.cs" />
+    <Compile Include="OITDemoViewModel.cs" />
     <Compile Include="ParticleViewModel.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
@@ -121,6 +122,9 @@
     <AppxManifest Include="Package.appxmanifest">
       <SubType>Designer</SubType>
     </AppxManifest>
+    <Content Include="NITRO_ENGINE.3ds">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
     <None Include="SimpleDemoW10_TemporaryKey.pfx" />
   </ItemGroup>
   <ItemGroup>
